bazel: update shard_count for test (#42491)
