+++ executor/jointest/BUILD.bazel
@@ -1,28 +1,24 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_test")
 
 go_test(
-    name = "fktest_test",
+    name = "jointest_test",
     timeout = "short",
     srcs = [
-        "foreign_key_test.go",
+        "join_test.go",
         "main_test.go",
     ],
     flaky = True,
-    shard_count = 25,
+    shard_count = 50,
     deps = [
         "//config",
-        "//ddl",
-        "//domain",
-        "//infoschema",
-        "//meta",
         "//meta/autoid",
-        "//parser/auth",
-        "//parser/model",
         "//planner/core",
-        "//sessiontxn",
+        "//session",
+        "//sessionctx/variable",
         "//testkit",
-        "//testkit/testsetup",
-        "//util/dbterror",
+        "//util",
+        "//util/dbterror/exeerrors",
+        "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_stretchr_testify//require",
         "@com_github_tikv_client_go_v2//tikv",
         "@org_uber_go_goleak//:goleak",
